client: further clarify external interfaces
